доработка страницы профиля, добавлены страницы заказов и профеля
