+++ src/pages/profile/profile-orders.js
@@ -1,0 +1,9 @@
+import styles from './profile.module.scss';
+
+function ProfileOrders() {
+  return (
+    <div className={`${styles.orders} mt-10`}>Cooming soon...</div>
+  )
+}
+
+export default ProfileOrders;
