+++ src/pages/profile/profile.js
@@ -1,24 +1,13 @@
-import { useEffect, useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
 import { Link, useHistory } from 'react-router-dom';
-import { Input, Button } from '@ya.praktikum/react-developer-burger-ui-components';
-import { getUser, patchUser, getLogout } from '../../services/actions/auth';
+import { getUser, getLogout } from '../../services/actions/auth';
 
 import styles from './profile.module.scss';
 
-function Profile() {
+function Profile({ children }) {
   const dispatch = useDispatch();
   const history = useHistory();
-  const { user } = useSelector(store => store.auth);
-  const [form, setValue] = useState({
-    email: user?.email ?? '',
-    password: user?.password ?? '',
-    name: user?.name ?? '',
-  })
-
-  const onChange = (e) => {
-    setValue({ ...form, [e.target.name]: e.target.value });
-  }
 
   useEffect(() => {
     dispatch(getUser());
@@ -57,47 +46,7 @@ function Profile() {
           </p>
         </div>
       </div>
-      <div className={styles.fields}>
-        <form onSubmit={() => { dispatch(patchUser({ name: form.name, email: form.email })) }}>
-          <Input
-            type='text'
-            placeholder='Имя'
-            onChange={onChange}
-            icon='EditIcon'
-            value={form.name}
-            name='name'
-          />
-          <Input
-            type='email'
-            placeholder='E-mail'
-            onChange={onChange}
-            icon='EditIcon'
-            value={form.email}
-            name='email'
-          />
-          <Input
-            type='password'
-            placeholder='Пароль'
-            onChange={onChange}
-            icon='EditIcon'
-            value={form.password || ''}
-            name='password'
-          />
-          <div className={styles['form-buttons']}>
-            <Button
-              type='primary'
-              htmlType='submit'
-              disabled={form?.name === user?.name && form?.email === user?.email && form?.password === ''}
-              size='medium'>Сохранить</Button>
-            <Button 
-            type='primary' 
-            onClick={() => { setValue(user) }} 
-            disabled={form?.name === user?.name && form?.email === user?.email && form?.password === ''}
-            htmlType='button' 
-            size='medium'>Отмена</Button>
-          </div>
-        </form>
-      </div>
+      {children}
     </div>
   )
 }
